nfs: move NFSVolume to shared types package
